revert: fix(compiler): corrects component transclusion on ...
This reverts commit 64a5ffe723942b1a5a25b7708a4f9715bc9a3325.

The original commit was fixing two issues - one of them was
preventing attributes that triggered directives that replaced
the compiled node to be merged into the new node.

This change was a breaking change (as seen in the diff of the
tests in this commit) and that's why it's being removed.

A proper fix will follow.