style(Angular.js): remove redundant `_angular`
Going through the commit history on GitHub, the `_angular` is for noConflict mode,
the case where you have an old reference to a variable called Angular you want to preserve.

Here is the commit that added _angular in : https://github.com/angular/angular.js/commit/
6ba010d8669e4d02c5ae15bed4fcf9ccb5dafca5

This feature was later removed here:
https://github.com/angular/angular.js/commit/1bb2805a79978f70d3b4f2c775d77f3300f03ea7

The variable 'made it through', it's redundant now.

Closes #7215