fix($compile): set the iteration state before linking
This issue was introduced in 824d370fd124d36f41c02a59ec840520fa2613c6.
The state for each row has to be set up *before* linking.

The cloneFn (the function passed into $transclude) is called *before* actual linking and thus it is enough to update the state inside the cloneFn callback.