chore(npm-shrinkwrap): fix karma dependencies
Updating to karma 0.12.13 (in commit e700af1235ac6bac5bfa9833267b8c6795b20511)
caused `iit` and `ddescribe` to crash and disconnect the browser stopping the
test run.

It appears that the problem is with one of the dependencies of karma rather
than karma itself. At least one of the karma dependencies updated in line
with karma's dependencies' semver specifications but subtly changed their
behaviour to break karma.  Possibly this is related to chokidar, glob,
minimatch or fsevents.