fix(ngAnimate): throw an error if a callback is passed to animate methods
As of b892b6e41f7494607de784b7f714b296f6a701dc (released in 1.3.0) it is no longer
valid to pass a callback to the following functions: `enter`, `move`, `leave`, `addClass`,
`removeClass`, `setClass` and `animate`.

To prevent confusing error messages, this change asserts that this parameter is
not a function.

Closes #11826
Closes #11713